Imported GNU Classpath gcj-import-20051117.        * gnu/java/net/protocol/file/Connection.java: Removed, fully merged.        * sources.am: Regenerated.        * Makefile.in: Likewise.
git-svn-id: svn+ssh://gcc.gnu.org/svn/gcc/trunk@107153 138bc75d-0d04-0410-961f-82ee72b054a4

+++ libjava/classpath/javax/swing/plaf/basic/BasicLookAndFeel.java
@@ -1159,7 +1159,7 @@ public abstract class BasicLookAndFeel extends LookAndFeel
       "Tree.hash", new ColorUIResource(new Color(128, 128, 128)),
       "Tree.leftChildIndent", new Integer(7),
       "Tree.rightChildIndent", new Integer(13),
-      "Tree.rowHeight", new Integer(16),
+      "Tree.rowHeight", new Integer(0),
       "Tree.scrollsOnExpand", Boolean.TRUE,
       "Tree.selectionBackground", new ColorUIResource(Color.black),
       "Tree.nonSelectionBackground", new ColorUIResource(new Color(255, 255, 255)),
